The goal is to copy the component styles to the public directory and reference the style files in the <head> tag.
Installing and importing dotenv

makeComponentCssToPublic.js - Copies the style files from the components folder to the public directory if NODE_ENV is set to development. It is executed at every server startup, after emptying the components folder in the public folder.

Importing makeComponentCssToPublic.js in 'server.js'

test.css - the original in the component folder, the copy in the public folder

head.ejs - access the headCssLinks property of the viewModel [array]

The render function merges the html parts belonging to the partial, if there are more, the function passes the link of the styles belonging to the partials in an array to head.ejs.

The paths of the styles are given a unique link with a version parameter due to cache busting.

'server' serves the public directory statically, the version parameter caused an error with the 'req.url' option, 'pathname' is ok.

'handlerHome' uses Object.assign() to modify the received empty viewModel object (now correctly).